requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
kafka-python==2.2.15
minio==7.2.18
pycparser==2.23
pycryptodome==3.23.0
typing_extensions==4.15.0
urllib3==2.5.0
